// generated.tsx/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  /** The time the document was created */
  createdAt: Scalars['DateTime'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Asset>;
  /** The file name */
  fileName: Scalars['String'];
  /** The file handle */
  handle: Scalars['String'];
  /** The height of the file */
  height?: Maybe<Scalars['Float']>;
  /** List of Asset versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID'];
  imageGPU: Array<Gpu>;
  imagesCPU: Array<Cpu>;
  imagesLaptop: Array<Laptop>;
  imagesMotherboard: Array<Motherboard>;
  imagesStorageEquipment: Array<StorageEquipment>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Asset>;
  /** The mime type of the file */
  mimeType?: Maybe<Scalars['String']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** The file size */
  size?: Maybe<Scalars['Float']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  /** Get the url for the asset with provided transformations applied. */
  url: Scalars['String'];
  /** The file width */
  width?: Maybe<Scalars['Float']>;
};
// ...
export type AssetConnection = {
  __typename?: 'AssetConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<AssetEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type DocumentVersion = {
  __typename?: 'DocumentVersion';
  createdAt: Scalars['DateTime'];
  data?: Maybe<Scalars['Json']>;
  id: Scalars['ID'];
  revision: Scalars['Int'];
  stage: Stage;
};
// ...
export type Laptop = Node & {
  __typename?: 'Laptop';
  brand?: Maybe<Scalars['String']>;
  /** The time the document was created */
  createdAt: Scalars['DateTime'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<RichText>;
  discountPercent?: Maybe<Scalars['Int']>;
  /** Get the document in other stages */
  documentInStages: Array<Laptop>;
  /** List of Laptop versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID'];
  images?: Maybe<Asset>;
  manufacturer?: Maybe<Scalars['String']>;
  manufacturerLink?: Maybe<Scalars['String']>;
  newProduct?: Maybe<Scalars['Boolean']>;
  onDiscount?: Maybe<Scalars['Boolean']>;
  price?: Maybe<Scalars['Int']>;
  promotion?: Maybe<Scalars['Boolean']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug?: Maybe<Scalars['String']>;
  /** System stage field */
  stage: Stage;
  stock?: Maybe<Scalars['Int']>;
  subtitle?: Maybe<Scalars['String']>;
  title: Scalars['String'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  warranty?: Maybe<Scalars['Int']>;
};
// ...
export enum Locale {
  /** System locale */
  En = 'en'
}
// ...
export type Motherboard = Node & {
  __typename?: 'Motherboard';
  brand?: Maybe<Scalars['String']>;
  /** The time the document was created */
  createdAt: Scalars['DateTime'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<RichText>;
  discountPercent?: Maybe<Scalars['Int']>;
  /** Get the document in other stages */
  documentInStages: Array<Motherboard>;
  /** List of Motherboard versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID'];
  images?: Maybe<Asset>;
  manufacturer?: Maybe<Scalars['String']>;
  manufacturerLink?: Maybe<Scalars['String']>;
  newProduct?: Maybe<Scalars['Boolean']>;
  onDiscount?: Maybe<Scalars['Boolean']>;
  price?: Maybe<Scalars['Int']>;
  promotion?: Maybe<Scalars['Boolean']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug?: Maybe<Scalars['String']>;
  /** System stage field */
  stage: Stage;
  stock?: Maybe<Scalars['Int']>;
  subtitle?: Maybe<Scalars['String']>;
  title: Scalars['String'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  warranty?: Maybe<Scalars['Int']>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Retrieve a single asset */
  asset?: Maybe<Asset>;
  /** Retrieve document version */
  assetVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple assets */
  assets: Array<Asset>;
  /** Retrieve multiple assets using the Relay connection interface */
  assetsConnection: AssetConnection;
  /** Retrieve a single cPU */
  cPU?: Maybe<Cpu>;
  /** Retrieve document version */
  cPUVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple cPUs */
  cPUs: Array<Cpu>;
  /** Retrieve multiple cPUs using the Relay connection interface */
  cPUsConnection: CpuConnection;
  /** Retrieve a single gPU */
  gPU?: Maybe<Gpu>;
  /** Retrieve document version */
  gPUVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple gPUs */
  gPUs: Array<Gpu>;
  /** Retrieve multiple gPUs using the Relay connection interface */
  gPUsConnection: GpuConnection;
  /** Retrieve a single laptop */
  laptop?: Maybe<Laptop>;
  /** Retrieve document version */
  laptopVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple laptops */
  laptops: Array<Laptop>;
  /** Retrieve multiple laptops using the Relay connection interface */
  laptopsConnection: LaptopConnection;
  /** Retrieve a single motherboard */
  motherboard?: Maybe<Motherboard>;
  /** Retrieve document version */
  motherboardVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple motherboards */
  motherboards: Array<Motherboard>;
  /** Retrieve multiple motherboards using the Relay connection interface */
  motherboardsConnection: MotherboardConnection;
  /** Fetches an object given its ID */
  node?: Maybe<Node>;
  /** Retrieve a single scheduledOperation */
  scheduledOperation?: Maybe<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations */
  scheduledOperations: Array<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations using the Relay connection interface */
  scheduledOperationsConnection: ScheduledOperationConnection;
  /** Retrieve a single scheduledRelease */
  scheduledRelease?: Maybe<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases */
  scheduledReleases: Array<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases using the Relay connection interface */
  scheduledReleasesConnection: ScheduledReleaseConnection;
  /** Retrieve a single storageEquipment */
  storageEquipment?: Maybe<StorageEquipment>;
  /** Retrieve document version */
  storageEquipmentVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple storage_Equipment */
  storage_Equipment: Array<StorageEquipment>;
  /** Retrieve multiple storage_Equipment using the Relay connection interface */
  storage_EquipmentConnection: StorageEquipmentConnection;
  /** Retrieve a single user */
  user?: Maybe<User>;
  /** Retrieve multiple users */
  users: Array<User>;
  /** Retrieve multiple users using the Relay connection interface */
  usersConnection: UserConnection;
};
// ...
export type RichText = {
  __typename?: 'RichText';
  /** Returns HTMl representation */
  html: Scalars['String'];
  /** Returns Markdown representation */
  markdown: Scalars['String'];
  /** Returns AST representation */
  raw: Scalars['RichTextAST'];
  /** Returns plain-text contents of RichText */
  text: Scalars['String'];
};
// ...
export type StorageEquipment = Node & {
  __typename?: 'StorageEquipment';
  brand?: Maybe<Scalars['String']>;
  /** The time the document was created */
  createdAt: Scalars['DateTime'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<RichText>;
  discountPercent?: Maybe<Scalars['Int']>;
  /** Get the document in other stages */
  documentInStages: Array<StorageEquipment>;
  /** List of StorageEquipment versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID'];
  images?: Maybe<Asset>;
  manufacturer?: Maybe<Scalars['String']>;
  manufacturerLink?: Maybe<Scalars['String']>;
  newProduct?: Maybe<Scalars['Boolean']>;
  onDiscount?: Maybe<Scalars['Boolean']>;
  price?: Maybe<Scalars['Int']>;
  promotion?: Maybe<Scalars['Boolean']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug?: Maybe<Scalars['String']>;
  /** System stage field */
  stage: Stage;
  stock?: Maybe<Scalars['Int']>;
  subtitle?: Maybe<Scalars['String']>;
  title: Scalars['String'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  warranty?: Maybe<Scalars['Int']>;
};
// ...
export type ProductDataQuery = { __typename?: 'Query', gPUs: Array<{ __typename?: 'GPU', id: string, onDiscount?: boolean | null, newProduct?: boolean | null, promotion?: boolean | null, stock?: number | null, brand?: string | null, price?: number | null, title: string, slug?: string | null, subtitle?: string | null, manufacturerLink?: string | null, warranty?: number | null, images?: { __typename?: 'Asset', url: string } | null, description?: { __typename?: 'RichText', raw: any } | null }>, cPUs: Array<{ __typename?: 'CPU', id: string, onDiscount?: boolean | null, newProduct?: boolean | null, promotion?: boolean | null, stock?: number | null, brand?: string | null, price?: number | null, title: string, slug?: string | null, subtitle?: string | null, manufacturerLink?: string | null, warranty?: number | null, images?: { __typename?: 'Asset', url: string } | null, description?: { __typename?: 'RichText', raw: any } | null }>, laptops: Array<{ __typename?: 'Laptop', id: string, onDiscount?: boolean | null, newProduct?: boolean | null, promotion?: boolean | null, stock?: number | null, brand?: string | null, price?: number | null, title: string, slug?: string | null, subtitle?: string | null, manufacturerLink?: string | null, warranty?: number | null, images?: { __typename?: 'Asset', url: string } | null, description?: { __typename?: 'RichText', raw: any } | null }>, motherboards: Array<{ __typename?: 'Motherboard', id: string, onDiscount?: boolean | null, newProduct?: boolean | null, promotion?: boolean | null, stock?: number | null, brand?: string | null, price?: number | null, title: string, slug?: string | null, subtitle?: string | null, manufacturerLink?: string | null, warranty?: number | null, images?: { __typename?: 'Asset', url: string } | null, description?: { __typename?: 'RichText', raw: any } | null }>, storage_Equipment: Array<{ __typename?: 'StorageEquipment', id: string, onDiscount?: boolean | null, newProduct?: boolean | null, promotion?: boolean | null, stock?: number | null, brand?: string | null, price?: number | null, title: string, slug?: string | null, subtitle?: string | null, manufacturerLink?: string | null, warranty?: number | null, images?: { __typename?: 'Asset', url: string } | null, description?: { __typename?: 'RichText', raw: any } | null }> };